Fix volatile type mismatch in hardware test get_speed()
Same cast fix as hw_esp32s3.h — ESP32 GCC 14 requires exact type match
for std::max template arguments.

+++ hardware_test_esp32s3/hardware_test_esp32s3.ino
@@ -347,7 +347,7 @@ void IRAM_ATTR taho_isr() {
 
 float get_speed() {
     unsigned long elapsed = (unsigned long)(micros() - taho_last);
-    elapsed = max(elapsed, taho_iv);
+    elapsed = max(elapsed, (unsigned long)taho_iv);
     if (taho_iv == 0 || elapsed > 500000UL) return 0.0f;
     return WHEEL_CIRC_M / (PULSES_PER_REV * (elapsed / 1e6f));
 }
